wifi: iwlwifi: mld: refactor setting key->hw_key_idx
This idx uses an indication whether the key was programed to the FW or
not.
Instead of setting it in all the special cases:
- igtk (from which mac80211 can have 2 but FW only one)
- Keys removed with iwl_mld_remove_ap_keys which still stays in mac80211
  for a while
- SW crypto keys

Set every key that mac80211 tries to add as INVALID (not in FW),
If it will be added, it will be marked as valid.

When a key is removed, always set it as INVALID and not only in the
special cases.

Upon restart, mark all keys as INVALID.

type=cleanup
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I8c1dc78bacbd3c17c4200ee481388efc578c38f2
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/155883
automatic-review: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 67ddfb425661b9952d737b94293180572d7515d1