wifi: iwlwifi: mld: implement canceling of notifications for specific objects
Since all the async notifications are handled under the wiphy lock, and
the wiphy lock is used for large flows, it can take very long for a
notification to be handled. Sometimes, when the notification finally
gets to run, the object it refers to no longer exist, and then we get a
warning, or in worst case, a new object with the same FW id was already
created, and then we run the notification on the wrong object.

To mitigate this, introduce an API to cancel notifications of an object
that is about to be removed:

1. A notification that is specific to an object should register to the
   async handlers infra providing and the object type and a cancel
   function. The cancel function will check that a specific notification
   instance should really be canceled (mostly by matching the object id).

2. Then, when the object is about to be removed, this API can be called
   with the right object type and id.

3. The API will iterate over all pending notifications, and run the
   cancel function for the ones with the right object type.

4. If the cancel function returns true, it will delete the notification
   from the list.

This implements the API and adds the macros to register a notification
to the async handling infra with and without the canceling option.

The actual use will be done in next patches.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I96dd559fee830396cae6a2691fc8242828be4340
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/156186
automatic-review: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 6aaac27994ddd5c693b9f0a24dc0ca30362e83d5