wifi: iwlwifi: cancel missed_beacon notif if the link is deactivated
If a link is being deactivated in FW, we shouldn't proccess any missed
beacon notification for it. After the link is deactivated, which means
that no more missed beacon notifications can come for that link, cancel
any pending missed beacon notification, so it'll not be processed after
the link is deactivated.

If the link which is being removed is the link indicated in
notif->other_link_id, we can still process the notification (for example
we can still disconnect). So simply mark the other link as invalid so
when the handler runs it'll not do anything with that.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I481bdd5fa983cfd5b24858b2475a7a0b5cfed1c8
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/156185
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 589853a485eb89e0307f43b7a123039e0ec2360e