wifi: iwlwifi: mld: cancel pending session protection notifs
After a link is being deactivated, we no longer expect to receive
session protection notifications for it.
However, such a notification could have been sent before the deactivation
and it is then stuck waiting for the wiphy lock (which it will not get
before the link is deactivated).
So when the notification is finally handled, the link is no longer
active, or doesn't exist, or a new link with the same id was already
created.
Anyway, we want to cancel that pending notifcation after the link is
being deactivated.
Use the cancellable-notification-handlers infra for that purpose.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ie6c6a75575e2446f262db2be12c222797868b670
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/156187
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 6e699f3e31245f978a2100a440febab701a05c1b