wifi: iwlwifi: mld: reorder notification handling
Notify the waiter of the notification (if there is one) only after the
notification was registered to the async handling (i.e. it was added to
async_handlers_list)/ or the sync handler was executed.

That way, the wait handler function doesn't have to validate the size
of the notification: it will be validated before adding it to
async_handlers_list/executing it (in case of a sync handler).
If it is not valid it'll be dropped and the waiter
will have a timeout.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I497d839ba75198339a4b6f9c30ebd6f39a53b6b7
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/162224
automatic-review: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 5cee21aa52ee4d3b92c53dd257c9f4e5637a2a6d