[BUGFIX] wifi: iwlwifi: mld: delete unneeded logic
iwl_mld_unblocked_emlsr is called when the last EMLSR blocker is
removed. It checks if the last exit from EMLSR was recent,
and if so - retruns immediately to EMLSR without link selection.

But this logic was replaced with the one in
iwl_mld_trigger_link_selection, so it is not needed. Remove it.

type=bugfix
ticket=none
fixes=If79b322ebfda6533f1140b8e6071d7d7198b1aa2

Change-Id: I9b1ff0bffb427bbe6f5a4361b591041e099ee093
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>x-iwlwifi-stack-dev: 0c0a5028e7b17e74d1b7aa2873f2c353337bc2d8