[BUGFIX] [NOUPSTREAM] wifi: iwlwifi: mld: Indicate SUPPORTS_AMSDU_IN_AMPDU only when needed
This flag inidcates to mac80211 if AMSDU in AMPDU is allowed, and it is
used to determine the value that will be set in the addba response.

In debug mode, AMSDU in AMPDU can be disabled. In this case, we shouldn't
turn this flag on.

type=bugfix
ticket=none
fixes=unknown

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Icdc5c75d52182ede274511b8708fea82703de84b
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/177303
x-iwlwifi-stack-dev: 15ef170d397b856b56fbfc7766c9c978ab4c0d69